requirements.txt
crewai>=0.11.0
python-dotenv>=1.0.0
langchain>=0.1.0
openai>=1.12.0
requests>=2.31.0
python-dateutil>=2.8.2 